refactor: Split tauri related components info platform specific
