+++ application/ui/src/api/client.ts
@@ -8,23 +8,7 @@ import createClient from 'openapi-react-query';
 
 import type { paths } from './openapi-spec';
 
-/* eslint-disable no-underscore-dangle */
-export const isTauriContext = (): boolean => typeof window.__TAURI__?.core?.invoke === 'function';
-
-const getBaseUrl = async (): Promise<string> => {
-    if (isTauriContext()) {
-        const tauriApiUrl = await window.__TAURI__!.core!.invoke<string>('get_public_api_url');
-
-        console.info('Backend public API URL:', tauriApiUrl);
-
-        return tauriApiUrl;
-    }
-
-    return import.meta.env.PUBLIC_API_URL || '';
-};
-/* eslint-enable no-underscore-dangle */
-
-export const baseUrl = await getBaseUrl();
+export const baseUrl = import.meta.env.PUBLIC_API_BASE_URL || '';
 
 export const client = createFetchClient<paths>({
     baseUrl,
