+++ application/ui/src/features/sources-sinks/sources/images-folder/images-folder-fields.component.tauri.tsx
@@ -3,7 +3,28 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Content, ContextualHelp, Flex, Heading, Text, TextField } from '@geti/ui';
+import { Button, Content, ContextualHelp, Flex, Heading, Text, TextField } from '@geti/ui';
+import { open } from '@tauri-apps/plugin-dialog';
+
+const getSingleSelectedPath = (selectedPath: string | string[] | null): string | null => {
+    if (typeof selectedPath === 'string') {
+        return selectedPath;
+    }
+
+    if (Array.isArray(selectedPath)) {
+        return selectedPath.length > 0 ? selectedPath[0] : null;
+    }
+    return null;
+};
+
+export const pickFolderPath = async (): Promise<string | null> => {
+    const selectedPath = await open({
+        directory: true,
+        multiple: false,
+    });
+
+    return getSingleSelectedPath(selectedPath);
+};
 
 const FolderPathDescription = () => {
     return (
@@ -27,6 +48,14 @@ interface ImagesFolderFieldsProps {
 }
 
 export const ImagesFolderFields = ({ folderPath, onSetFolderPath }: ImagesFolderFieldsProps) => {
+    const handleBrowse = async (): Promise<void> => {
+        const selectedPath = await pickFolderPath();
+
+        if (selectedPath !== null) {
+            onSetFolderPath(selectedPath);
+        }
+    };
+
     return (
         <Flex alignItems={'end'} gap={'size-100'}>
             <TextField
@@ -37,6 +66,9 @@ export const ImagesFolderFields = ({ folderPath, onSetFolderPath }: ImagesFolder
                 contextualHelp={<FolderPathDescription />}
                 isRequired
             />
+            <Button variant={'secondary'} onPress={handleBrowse}>
+                Browse
+            </Button>
         </Flex>
     );
 };
